@@ -117,52 +124,53 @@
     "@types/classnames": "^2.2.10",
     "@types/express": "^4.17.6",
     "@types/history": "^4.7.6",
-    "@types/jest": "^25.2.3",
-    "@types/lodash": "^4.14.152",
+    "@types/jest": "^26.0.0",
+    "@types/lodash": "^4.14.155",
     "@types/qs": "^6.9.3",
-    "@types/react": "^16.9.35",
+    "@types/react": "^16.9.36",
     "@types/react-document-title": "^2.0.4",
     "@types/react-dom": "^16.9.8",
     "@types/react-helmet": "^6.0.0",
-    "@umijs/fabric": "^2.0.8",
-    "@umijs/plugin-blocks": "^2.1.3",
+    "@typescript-eslint/eslint-plugin": "^3.2.0",
+    "@umijs/fabric": "^2.1.0",
+    "@umijs/plugin-blocks": "^2.1.4",
     "@umijs/preset-ant-design-pro": "^1.2.2",
-    "@umijs/preset-react": "^1.5.3",
-    "@umijs/preset-ui": "^2.1.14",
-    "@typescript-eslint/eslint-plugin": "^3.0.0",
+    "@umijs/preset-react": "^1.5.6",
+    "@umijs/preset-ui": "^2.1.15",
     "carlo": "^0.9.46",
-    "chalk": "^4.0.0",
+    "chalk": "^4.1.0",
     "check-prettier": "^1.0.3",
     "cross-env": "^7.0.2",
     "cross-port-killer": "^1.2.1",
     "detect-installer": "^1.0.2",
     "enzyme": "^3.11.0",
-    "eslint": "^7.1.0",
-    "eslint-plugin-eslint-comments": "^3.1.2",
-    "eslint-plugin-jest": "^23.13.1",
-    "eslint-plugin-unicorn": "^20.0.0",
+    "eslint": "^7.2.0",
+    "eslint-plugin-eslint-comments": "^3.2.0",
+    "eslint-plugin-jest": "^23.13.2",
+    "eslint-plugin-unicorn": "^20.1.0",
     "express": "^4.17.1",
-    "gh-pages": "^2.2.0",
+    "gh-pages": "^3.0.0",
     "husky": "^4.2.5",
     "import-sort-cli": "^6.0.0",
     "import-sort-parser-babylon": "^6.0.0",
     "import-sort-parser-typescript": "^6.0.0",
     "import-sort-style-module": "^6.0.0",
     "jest-puppeteer": "^4.4.0",
-    "lint-staged": "^10.2.6",
+    "jsdom-global": "^3.0.2",
+    "lint-staged": "^10.2.10",
     "mockjs": "^1.1.0",
     "node-fetch": "^2.6.0",
-    "npm-check-updates": "^6.0.1",
+    "npm-check-updates": "^7.0.1",
     "prettier": "^2.0.5",
     "pro-download": "1.0.1",
-    "puppeteer-core": "^3.1.0",
+    "puppeteer-core": "^3.3.0",
     "slash2": "^2.0.0",
-    "stylelint": "^13.5.0",
+    "stylelint": "^13.6.0",
     "stylelint-config-prettier": "^8.0.1",
     "stylelint-config-standard": "^20.0.0"
   },
   "optionalDependencies": {
-    "puppeteer": "^3.1.0"
+    "puppeteer": "^3.3.0"
   },
   "engines": {
     "node": ">=10.0.0"
