@@ -1,13 +1,16 @@
 // 此处放置其他初始化代码
 
-window.appInitCustom = {
-  platformName: 'Code构建工具',
-  appName: 'Code构建工具',
+window.appInitCustomData = {
+  platformName: '亲优亲选',
+  appName: '亲优亲选 城市运营商管理系统',
+  appDescription: '摒弃虚浮，潜心笃志，坚持以技术为系统建设核心',
   loginLogo: '/logo.png',
   shareLogo: '/shareLogo.png',
-  shareLogoName: 'Code构建工具',
-  companyName: '×××科技有限公司',
+  shareLogoName: '城市运营商管理系统',
+  leftBarText: '亲优亲选商城',
+  companyName: '潘哆拉科技有限公司',
+  copyright: '2020 河南潘哆拉科技有限公司 技术部出品',
   apiPrefix: {
-    corsTargetProduction: '',
+    corsTargetProduction: 'https://api2.yurukeji.com.cn',
   },
 };
