@@ -7,7 +7,7 @@ import CheckPermissions from './CheckPermissions';
 
 const Exception403 = () => 403;
 
-export const isComponentClass = component => {
+export const isComponentClass = (component) => {
   if (!component) return false;
   const proto = Object.getPrototypeOf(component);
   if (proto === React.Component || proto === Function.prototype) return true;
@@ -17,14 +17,14 @@ export const isComponentClass = component => {
 // Authorized  render is already instantiated, children is no instantiated
 // Secured is not instantiated
 
-const checkIsInstantiation = target => {
+const checkIsInstantiation = (target) => {
   if (isComponentClass(target)) {
     const Target = target;
-    return props => <Target {...props} />;
+    return (props) => <Target {...props} />;
   }
 
   if (React.isValidElement(target)) {
-    return props => React.cloneElement(target, props);
+    return (props) => React.cloneElement(target, props);
   }
 
   return () => target;
