@@ -1,8 +1,10 @@
 import React, { PureComponent } from 'react';
-import { formatMessage, setLocale, getLocale } from 'umi';
+import { setLocale, getLocale } from 'umi';
 import classNames from 'classnames';
 import { Menu, Dropdown } from 'antd';
 import { GlobalOutlined } from '@ant-design/icons';
+
+import { formatMessage } from '@/utils/tools';
 
 import styles from './index.less';
 
@@ -14,6 +16,7 @@ export default class SelectLang extends PureComponent {
   render() {
     const { className } = this.props;
     const selectedLang = getLocale();
+
     const langMenu = (
       <Menu className={styles.menu} selectedKeys={[selectedLang]} onClick={this.changLang}>
         <Menu.Item key="zh-CN">
