@@ -1,8 +1,8 @@
 import { Button, message, notification } from 'antd';
 import React from 'react';
-import { formatMessage } from 'umi';
 
-import defaultSettings from '@/defaultSettings';
+import { formatMessage } from '@/utils/tools';
+import { defaultSettings } from '@/defaultSettings';
 
 const { pwa } = defaultSettings; // if pwa is true
 
