@@ -1,10 +1,12 @@
-import { DefaultFooter, getMenuData, getPageTitle } from '@ant-design/pro-layout';
-import { Helmet, HelmetProvider } from 'react-helmet-async';
-import { connect, Link, formatMessage } from 'umi';
 import React from 'react';
+import { connect, Link } from 'umi';
+import { Helmet, HelmetProvider } from 'react-helmet-async';
+import { DefaultFooter, getMenuData, getPageTitle } from '@ant-design/pro-layout';
 
+import { formatMessage } from '@/utils/tools';
 import SelectLang from '@/components/SelectLang';
-import logo from '@/assets/logo.svg';
+import { defaultSettings } from '@/defaultSettings';
+import { showSelectLanguage, showLogoInLoginView, appInitCustom } from '@/customConfig/config';
 
 import styles from './UserLayout.less';
 
@@ -22,6 +24,7 @@ const UserLayout = (props) => {
     },
   } = props;
   const { breadcrumb } = getMenuData(routes);
+
   const title = getPageTitle({
     pathname: location.pathname,
     formatMessage,
@@ -36,22 +39,29 @@ const UserLayout = (props) => {
       </Helmet>
 
       <div className={styles.container}>
-        <div className={styles.lang}>
-          <SelectLang />
-        </div>
+        <div className={styles.lang}>{showSelectLanguage() ? <SelectLang /> : null}</div>
         <div className={styles.content}>
           <div className={styles.top}>
             <div className={styles.header}>
               <Link to="/">
-                <img alt="logo" className={styles.logo} src={logo} />
-                <span className={styles.title}>Ant Design</span>
+                {showLogoInLoginView() ? (
+                  <img alt="logo" className={styles.logo} src={defaultSettings.getShareLogo()} />
+                ) : null}
+                <span className={styles.title}>
+                  {appInitCustom == null ? '未设置名称' : appInitCustom.appName || '未设置名称'}
+                </span>
               </Link>
             </div>
-            <div className={styles.desc}>Ant Design 是西湖区最具影响力的 Web 设计规范</div>
+            <div className={styles.desc}>
+              {appInitCustom == null ? '' : appInitCustom.appDescription || ''}
+            </div>
           </div>
           {children}
         </div>
-        <DefaultFooter />
+        <DefaultFooter
+          links={[]}
+          copyright={appInitCustom == null ? '' : appInitCustom.copyright || ''}
+        />
       </div>
     </HelmetProvider>
   );
