@@ -1,8 +1,27 @@
 import React from 'react';
-import { connect, Redirect } from 'umi';
+import Redirect from 'umi/redirect';
+import { connect } from 'umi';
+import { pathToRegexp } from 'path-to-regexp';
 import Authorized from '@/utils/Authorized';
-import { getRouteAuthority } from '@/utils/utils';
 import { getToken } from '@/utils/tools';
+
+const getRouteAuthority = (path, routeData) => {
+  let authorities;
+  routeData.forEach((route) => {
+    // match prefix
+    if (pathToRegexp(`${route.path}(.*)`).test(path)) {
+      // exact match
+      if (route.path === path) {
+        authorities = route.authority || authorities;
+      } // get children authority recursively
+
+      if (route.routes) {
+        authorities = getRouteAuthority(path, route.routes) || authorities;
+      }
+    }
+  });
+  return authorities;
+};
 
 const AuthComponent = ({
   children,
